Formatted the code a bit and remove a panic
